requirements.txt
telethon
python-dotenv
pyquotex @ git+https://github.com/cleitonleonel/pyquotex.git